% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce_columns}
\alias{coalesce_columns}
\title{Coalesce columns into a single new column}
\usage{
coalesce_columns(
  x,
  new_var,
  var_list,
  remove = TRUE,
  drop = FALSE,
  warn_only = TRUE,
  pick_first = TRUE
)
}
\arguments{
\item{var_list}{}
}
\description{
Coalesce columns into a single new column
}
\note{
The default operation of this function should mirror \code{\link[dplyr:coalesce]{dplyr::coalesce()}}.
}
