% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_columns.R
\name{detect_duplicate_colnames}
\alias{detect_duplicate_colnames}
\title{Detect duplicate colnames in a data frame}
\usage{
detect_duplicate_colnames(.x)
}
\arguments{
\item{.x}{A data frame}
}
\value{
A vector of named logical values indicating if colname
is duplicated in data frame.
}
\description{
Detect duplicate colnames in a data frame
}
\examples{
\dontrun{
detect_duplicate_colnames(iris)
# Sepal.Length  Sepal.Width Petal.Length  Petal.Width      Species
# FALSE        FALSE        FALSE        FALSE        FALSE
}
}
