% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyrx.R
\docType{package}
\name{dplyrx-package}
\alias{dplyrx}
\alias{dplyrx-package}
\title{dplyrx: an extension of dplyr}
\description{
dplyrx provides a collection of functions to work on data frames that
complement those of the dplyr package.
}
\details{
To learn more about dplyrx, start with the vignettes:
`browseVignettes(package = "dplyrx")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/choisy/gamar}
  \item Report bugs at \url{https://github.com/choisy/gamar/issues}
}

}
\author{
\strong{Maintainer}: Marc Choisy \email{marc.choisy@ird.fr} (0000-0002-5187-6390) [copyright holder]

Authors:
\itemize{
  \item Lucie Contamin \email{contamin.lucie@gmail.com} (0000-0001-5797-1279)
}

}
