% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_columns.R
\name{find_common_colnames}
\alias{find_common_colnames}
\title{Find common colnames between data frames}
\usage{
find_common_colnames(tbls = NULL, ignore = NULL)
}
\arguments{
\item{ignore}{A list of column names to ignore when finding common names}

\item{...}{A list of data.frames}
}
\value{
A character vector of column names that are in common between
data frame.
}
\description{
Given data.frames (as parameters), check what column names
are in common among them.
}
\examples{
\dontrun{
find_common_colnames(iris,iris)
#[1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"
}
}
