% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{left_join}
\alias{left_join}
\alias{right_join}
\alias{inner_join}
\alias{full_join}
\title{Left join with coalesce of column collisions}
\usage{
left_join(x, y, by = NULL, suffix = c(".x", ".y"), ...)

right_join(x, y, by = NULL, suffix = c(".x", ".y"), ...)

inner_join(x, y, by = NULL, suffix = c(".x", ".y"), ...)

full_join(x, y, by = NULL, suffix = c(".x", ".y"), ...)
}
\arguments{
\item{x}{The first table}

\item{y}{The second table}

\item{by}{A join variable (see \code{\link[dplyr:mutate-joins]{dplyr::left_join()}})}

\item{suffix}{A two-element list of suffixes to append if colnames collide.}

\item{...}{Any other parameters (see \code{\link[dplyr:mutate-joins]{dplyr::left_join()}})}
}
\value{
A data frame representing the join of the two parameters, followed by an attempt
to coalesce columns that collided (same column name). If they cannot be coalesced, they
are left as is with a warning printed.
}
\description{
When joining two tables, collisions can occur and the column names will
be appended with a suffix. This function then tries to re-combine these
columns based on equality or NA.
}
\section{Functions}{
\itemize{
\item \code{right_join()}: Right join

\item \code{inner_join()}: Inner join

\item \code{full_join()}: Full join

}}
