% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{list_join}
\alias{list_join}
\alias{left_joinl}
\alias{right_joinl}
\alias{inner_joinl}
\alias{full_joinl}
\title{Join a list of tables together on a common key}
\usage{
list_join(x, by = NULL, which = "full", suffix = NULL, ...)

left_joinl(...)

right_joinl(...)

inner_joinl(...)

full_joinl(...)
}
\arguments{
\item{x}{A list of data frames}

\item{by}{A join key (must be common across all tables)}

\item{which}{The type of join (left, right, inner, full). See \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}.}

\item{suffix}{A list of suffixes to add to column names that collide between tables.}

\item{...}{Any other parameters for the join function.}
}
\value{
A data frame with all tables joined on \code{by}
}
\description{
Join a list of tables using \code{which} join method, by a common
\code{key} value. If variables exist in common across tables, the \code{suffix} list
will be appended to the column name. The result is a joined table from across
all list elements of \code{x}.
}
\section{Functions}{
\itemize{
\item \code{left_joinl()}: Include all rows from cumulatively joined table

\item \code{right_joinl()}: Include all rows from each additional table

\item \code{inner_joinl()}: Include rows in common between all tables

\item \code{full_joinl()}: Include rows in any table

}}
\examples{
\dontrun{
list_join(list(iris,iris), by="Species", which="full", suffix=c(".1",".2"))
}
}
