% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo}
\alias{meteo}
\title{Meteoroligical data for 4 stations and 3 years in Vietnam.}
\format{A data frame with 36 rows and 7 variables:
\describe{
  \item{year}{the year of collection.}
  \item{month}{the month of collection.}
  \item{station}{the name of the climatic station.}
  \item{temperature}{monthly average temperature in Celcuis degrees.}
  \item{humidity}{monthly average absolute humidity in g / m^3.}
  \item{rainfall}{monthly cumulative rainfall in mm.}
  \item{sunshine}{monthly cumulative number of hours of sunshine.}
}}
\source{
The Vietnamese Institute of Meteorology, Hydrology and Climate
Change: \url{http://www.imh.ac.vn}.
}
\usage{
meteo
}
\description{
A dataset containing the temperature, the humidity, the cumulative rainfall
and the number of hours of sunshine for 4 climatic stations of Vietnam for
the months of January, May and September of the years 2013, 2014 and 2015.
}
\keyword{datasets}
