% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_by.R
\name{mutate_by}
\alias{mutate_by}
\title{Mutate one column value by summarize value}
\usage{
mutate_by(df, .filter, .funs, ..., colgroups = NULL)
}
\arguments{
\item{df}{the data frame on which to perform the mutate_by.}

\item{.filter}{a filtering function to select the rows to applied the
mutate_by.}

\item{.funs}{function called to calculate the value to applied to the filter
value. Inputed as a function name.}

\item{...}{additional arguments passed on to function.}

\item{colgroups}{other variables to apply \code{group_by} used to calculate
the new value.}
}
\value{
A data frame with the same variables as \code{df}.
}
\description{
\code{mutate_by} mutates value of a certain variable by the summarize value
of the variable. If necessary, the summarize value can be calculate on
categorical variable groups
}
\details{
The function performs a \code{\link[dplyr]{filter}}, if necessary a
\code{\link[dplyr]{group_by}}, followed by a \code{\link[dplyr]{summarise}}
 with the function(s) \code{.funs}.
}
\examples{
library(dplyr)

## A toy data frame:
set.seed(30101976)
data <- expand.grid(letters[1:3], 1:3, 4:6)
data$Var4 <- sample(1:100, nrow(data), TRUE)
data$Var5 <- sample(1:100, nrow(data), TRUE)
data$Var6 <- sample(1:100, nrow(data), TRUE)

# To mutate the value > 75 in the variable `Var4` by the mean of the value of
# `Var4` except the one > 75.
data \%>\% mutate_by(Var4 > 75, mean)

# To mutate the value > 75 in the variable `Var4` by the mean of the value of
# `Var4` group by `Var1` and `Var2`, except the one > 75.
data \%>\% mutate_by(Var4 > 75, mean, colgroups = c("Var1", "Var2"))

# To mutate the value > 75 in the variable `Var4` by the mean of the value of
# `Var4` group by `Var1` and `Var2`, except the one > 75 and by removing NA
# value.
data \%>\% mutate_by(Var4 > 75, mean, colgroups = c("Var1", "Var2"),
  na.rm = TRUE)
}
\author{
Lucie Contamin.
}
