% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_columns.R
\name{rename_common_columns}
\alias{rename_common_columns}
\title{Rename potential collisions in column names for a list of data frames}
\usage{
rename_common_columns(tbls, suffix = NULL, ignore = NULL)
}
\arguments{
\item{tbls}{A list of tables}

\item{suffix}{A list of suffixes to add to colnames of a table (if a collision
occurs). Default is c(".t1",".t2",...)}

\item{ignore}{A list of colnames to ignore (e.g., keys for later joining)}
}
\value{
A list of tables with the colnames appended with suffixes where needed
to avoid collisions.
}
\description{
When considering a list of data frames, some of the column
names may be identical. If combining the data frames, there will be
collisions (same name). This function takes a list and add suffixes to
potential collision column names, to avoid this problem.
}
\examples{
\dontrun{
colnames(rename_collisions(list(iris, iris), suffix=c(".i1",".i2"))[[1]])
# [1] "Sepal.Length.i1" "Sepal.Width.i1"  "Petal.Length.i1" "Petal.Width.i1"
# [5] "Species.i1"
}
}
