% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by.R
\name{aggregate_by}
\alias{aggregate_by}
\title{Aggregating observations on some variables}
\usage{
aggregate_by(df, col_name, ..., .funs = sum)
}
\arguments{
\item{df}{the data frame on which to perform the aggregation.}

\item{col_name}{the variable on which to perform the aggregation}

\item{...}{other  variables on which to perform the aggregation}

\item{.sfun}{the function used to perform the aggregation.
By default \code{sum}.}
}
\value{
A data frame with the same variables as `df` but for which some of
the observation have been aggregated (i.e. less rows than in `df`).
}
\description{
\code{aggregate_by} aggregates some variables of a data frame according to
the values of a categorical variable of the same data frame and using a given
function for the aggregation.
}
\details{
On a data frame that contains the values of the categorical variable to
aggregate (`col_name`, `...`), the function performs a
\code{\link[dplyr]{group_by}} followed by a \code{\link[dplyr]{summarise}}
with the function(s) \code{.funs}.
}
\examples{
## A toy data frame:
set.seed(30101976)
data <- expand.grid(letters[1:3], 1:3, 4:6)
data$Var4 <- sample(1:100, nrow(data), TRUE)
data$Var5 <- sample(1:100, nrow(data), TRUE)
data$Var6 <- sample(1:100, nrow(data), TRUE)

## Aggregating the values "a" and "b" of the categorical variable Var1,
## summing the values of variables Var4, Var5, Var6 (i.e. all the variables
## that are not in the arguments of the function call):
aggregate_by(data, Var1, Var2, Var3)

}
\author{
Marc Choisy and Lucie Contamin.
}
