% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisions.R
\name{coalesce_collisions}
\alias{coalesce_collisions}
\title{Coalesce columns that collided if possible}
\usage{
coalesce_collisions(x, suffix = c(".x", ".y"))
}
\arguments{
\item{x}{A data frame with collisions}

\item{suffix}{Suffix values expected for collisions (not all suffixes need to be
in the collisions).}
}
\value{
A data frame with all possible collision columns coalesced. Remaining
columns that collided are kept as is.
}
\description{
When joining two data frames, columns of the same name will
be renamed as \code{.x} and \code{.y}. These collisions can be coalesced into a
single column provided they do not conflict. This function performs that
operation.
}
